requirements.txt
folium==0.15.1
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
Pillow==9.0.1
Pillow==10.2.0
plotly==5.19.0
streamlit==1.31.1
streamlit_folium==0.18.0
